+++ core/build.ts
@@ -87,11 +87,10 @@ export async function getOutput(destFile: DestFile, title: string, site: Site, r
 	}
 }
 
-export async function buildDest(site: Site, rt: Runtime) {
+export async function buildDest(site: Site, rt: Runtime, modeDryRun?: boolean) {
 	const config = site.config;
 	const destFiles = site.destFiles;
 	const destDir = config.dest!;
-	const modeDryRun = rt.options.dry_run!;
 
 	async function sub(filepathDir: string, pathDir: string) {
 		for await (const entry of Deno.readDir(filepathDir)) {